poprawki datagrid, konieczny wybor przy usuwaniu

+++ CompanyMaster/SearchPage.xaml.cs
@@ -54,20 +54,27 @@ namespace CompanyMaster
 
         private void deleteButton_Click(object sender, RoutedEventArgs e)
         {
-            CompaniesDatabaseEntities db = new CompaniesDatabaseEntities();
-            int CId = (companyDataGrid.SelectedItem as Companies).Id;
-            string CName = (companyDataGrid.SelectedItem as Companies).FullName;
-            
+            if (companyDataGrid.SelectedIndex != -1)
+            {
+                CompaniesDatabaseEntities db = new CompaniesDatabaseEntities();
+                int CId = (companyDataGrid.SelectedItem as Companies).Id;
+                string CName = (companyDataGrid.SelectedItem as Companies).FullName;
 
-            Companies comp = (from r in db.Companies where r.Id == CId select r).SingleOrDefault();
-            db.Companies.Remove(comp);
-            db.SaveChanges();
 
-            string xd = CId.ToString();
-            MessageBox.Show("Removed "+xd+" "+CName);
+                Companies comp = (from r in db.Companies where r.Id == CId select r).SingleOrDefault();
+                db.Companies.Remove(comp);
+                db.SaveChanges();
 
-            companyDataGrid.ItemsSource = db.Companies.ToList();
+                string xd = CId.ToString();
+                MessageBox.Show("Removed " + xd + " " + CName);
 
+
+                companyDataGrid.ItemsSource = db.Companies.ToList();
+            }
+            else
+            {
+                MessageBox.Show("Select record");
+            }
         }
     }
 }
